+++ src/mir_core/src/Windows/CCtrlClc.cpp
@@ -117,9 +117,9 @@ int CCtrlClc::GetExtraColumns() const
 {	return SendMessage(m_hwnd, CLM_GETEXTRACOLUMNS, 0, 0);
 }
 
-uint8_t CCtrlClc::GetExtraImage(HANDLE hItem, int iColumn) const
+uint16_t CCtrlClc::GetExtraImage(HANDLE hItem, int iColumn) const
 {
-	return (uint8_t)(SendMessage(m_hwnd, CLM_GETEXTRAIMAGE, (WPARAM)hItem, MAKELPARAM(iColumn, 0)) & 0xFFFF);
+	return (uint16_t)(SendMessage(m_hwnd, CLM_GETEXTRAIMAGE, (WPARAM)hItem, MAKELPARAM(iColumn, 0)) & 0xFFFF);
 }
 
 HIMAGELIST CCtrlClc::GetExtraImageList() const
